It's there: Port M3G to GLES 2.0!
Stea.. umm I mean "integrate" the M3G implementation from KEmulator, then rewrite it to use GLES 2.0. This was a lot of work, but hell, it's finally here!

This implementation might be a little less reliable than the reference Nokia one, but ultimately 3D engines shouldn't be implemented inside native code and the fixed pipeline can't be reliably used anymore, hence from now on we should be working with this implementation.. As a bonus, after porting we now have a proper two-sided lighting support.

The code is.. what it is..